+++ services/dashboard-ui/client/components/webhooks/CreateWebhook/index.ts
@@ -1,2 +1,1 @@
 export { CreateWebhookButton } from './CreateWebhookContainer'
-export { CreateWebhookModal as CreateWebhookModalComponent } from './CreateWebhook'
