+++ services/dashboard-ui/client/components/webhooks/EditWebhook/index.ts
@@ -1,2 +1,1 @@
 export { EditWebhookButton } from './EditWebhookContainer'
-export { EditWebhookModal as EditWebhookModalComponent } from './EditWebhook'
