+++ services/dashboard-ui/client/components/webhooks/WebhookForm/index.ts
@@ -1,0 +1,1 @@
+export { WebhookFormModal } from './WebhookForm'
